implement errors table

+++ Application/MainLogic/Inc/control_board.h
@@ -2,6 +2,7 @@
 #define __CONTROL_BOARD_H
 
 #include "app_state_machine.h"
+#include "error_codes.h"
 
 /**
  * @brief Main structure for ControlBoard-4wcl
@@ -10,8 +11,9 @@ typedef struct {
     SystemState_t current_state;
     uint32_t heartbeat_count;
     /* Add other principal variables here */
-    uint32_t error_flags;
+    uint64_t error_flags;
 } ControlBoard_t;
+
 
 extern ControlBoard_t ControlBoard_4wcl;
